+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.2
+
+* Limit image to a crop area instead of view boundaries
+* Don't adjust a size during scale to avoid misalignment
+* After editing snap image back to a crop area. Auto scale if needed
+
 ## 0.1.1
 
 * Fixed an exception when aspect ratio is not supplied